Fixed a gcc warning.

+++ storage/falcon/MemMgr.cpp
@@ -636,7 +636,7 @@ void MemMgr::releaseBlock(MemBlock *block)
 	sync.lock(Exclusive);
 	//validateFreeList();
 	block->pool = NULL;
-	ASSERT(length <= activeMemory);
+	ASSERT(length <= (int) activeMemory);
 	activeMemory -= length;
 	
 	if (freeBlock->next && !freeBlock->next->memHeader.pool)
